Full implementation of the key-value system.
Secrets are made available as folders, with their data and keys exposed as
files.

Secrets are currently loaded into local process memory and cached until a
Lookup() takes place on the directory of the secret.

This may not be the best way to do it, but its not clear how our relationship
to `lease_duration` should take place.