Refactor to support policy-limited path traversal.
Inaccessible keys are returned as empty-directories which support doing
additional lookups for subkeys. This allows policy which would deny access
to still allow directory traversal to subkeys.

The protocol with Vault is made much more chatty to ensure that the returned
paths always reflect the current state of the Vault server.